Initial commit: Replaces bvh function to snap nearest.
BVHtree to snap to vertices and edges is independent of ray. The function for previous transverse was unclear and not very efficient.
The function used now, projects four vertices of aabb to test if it is within the threshold.
It still needs to optimize even more (prioritize axis, anticipate return)

I'll still test the optimization